+++ rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust"
+name = "tsunagi_sdk"
 version = "0.1.0"
 edition = "2021"
